Collapse state_database mod
This change collapses the state_database module up into the root
state module, including the merkle_state module.
