+++ services/scabbard/libscabbard/src/service/state/merkle_state.rs
